sha3: move `xor_block`
