+++ sha3/src/block_api.rs
@@ -1,4 +1,4 @@
-use crate::{DEFAULT_ROUND_COUNT, PLEN, Sha3ReaderCore, xor_block};
+use crate::{DEFAULT_ROUND_COUNT, PLEN, Sha3ReaderCore};
 use core::{fmt, marker::PhantomData};
 use digest::{
     HashMarker, Output,
@@ -259,5 +259,22 @@ where
             state,
             _pd: PhantomData,
         })
+    }
+}
+
+pub(crate) fn xor_block(state: &mut [u64; PLEN], block: &[u8]) {
+    assert!(block.len() < 8 * PLEN);
+
+    let mut chunks = block.chunks_exact(8);
+    for (s, chunk) in state.iter_mut().zip(&mut chunks) {
+        *s ^= u64::from_le_bytes(chunk.try_into().unwrap());
+    }
+
+    let rem = chunks.remainder();
+    if !rem.is_empty() {
+        let mut buf = [0u8; 8];
+        buf[..rem.len()].copy_from_slice(rem);
+        let n = block.len() / 8;
+        state[n] ^= u64::from_le_bytes(buf);
     }
 }
